# openapi.yaml
openapi: 3.1.0
info:
  title: WinMyDispute API
  version: 3.1.0
  description: |
    GPT-facing API for the WinMyDispute free-preview and premium-upgrade flow.
    The free tier returns a likely reason-code match, a confidence read, and exactly one strategy tip.
    Premium output unlocks after a one-time Stripe Checkout payment tied to the user's email and can produce
    a professional dispute packet as JSON plus downloadable PDF, Word-compatible RTF, or text report output.
  contact:
    name: Daniel Neville
    email: pureplays@icloud.com
    url: https://pure-plays.com
  license:
    name: Proprietary
    url: https://pure-plays.com/terms
servers:
  - url: https://api.pure-plays.com
    description: Production custom domain
  - url: https://winmydispute-api.onrender.com
    description: Render deployment
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/denials/respond:
    post:
      tags: [disputes]
      summary: Generate a follow-up response after a denial
      description: |
        Requires a paid email license plus either a signed premium access token or the matching
        Stripe checkout session ID. Builds a reconsideration package after an issuer or merchant denial.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DisputeInput"
                - type: object
                  required: [email]
                  properties:
                    denialSummary:
                      type: string
                    denialReason:
                      type: string
                    issuerFinding:
                      type: string
                    merchantResponseText:
                      type: string
                    denialEvidence:
                      type: string
                    issuerReferenceNumber:
                      type: string
      responses:
        "200":
          description: Denial response package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DenialResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "402":
          description: Upgrade required before denial-response generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeRequiredResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/v1/evidence/extract:
    post:
      tags: [disputes]
      summary: Extract structured evidence from uploaded files
      description: |
        Premium endpoint for evidence ingestion. Requires a paid email license plus either a signed
        premium access token or the matching Stripe checkout session ID. Accepts multipart uploads and returns structured
        evidence items, timeline entries, rebuttal concerns, extracted text, and a combined summary.
        Text-like files are parsed directly, PDFs are parsed for embedded text when possible, and
        image/PDF OCR uses the configured OCR provider when available. Uploaded files are also turned
        into a labeled exhibit packet for downstream report generation.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [email, files]
              properties:
                email:
                  type: string
                  format: email
                premiumAccessToken:
                  type: string
                checkoutSessionId:
                  type: string
                sessionId:
                  type: string
                description:
                  type: string
                merchantName:
                  type: string
                issuer:
                  type: string
                network:
                  type: string
                async:
                  type: boolean
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Structured evidence extraction response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceExtractionResponse"
        "202":
          description: Evidence extraction queued asynchronously
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAcceptedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "402":
          description: Upgrade required before evidence extraction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeRequiredResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    BadRequest:
      description: Invalid request input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Request limit exceeded for the current rate-limit window
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RateLimitError"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: A valid email address is required.
        requestId:
          type: string
          format: uuid
          nullable: true
      required: [error]
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          example: Too many requests
        requestId:
          type: string
          format: uuid
          nullable: true
        retryAfterSeconds:
          type: integer
          minimum: 1
          example: 42
      required: [error, requestId, retryAfterSeconds]
    EvidenceItem:
      oneOf:
        - type: string
        - type: object
          additionalProperties: true
          properties:
            summary:
              type: string
            description:
              type: string
            text:
              type: string
            filename:
              type: string
            fileType:
              type: string
            type:
              type: string
            extractedText:
              type: string
    DisputeInput:
      type: object
      additionalProperties: true
      properties:
        email:
          type: string
          format: email
          description: Optional for preview, required for premium generation and license lookup.
        premiumAccessToken:
          type: string
          description: Signed premium access token returned by /auth/check-license after payment.
        checkoutSessionId:
          type: string
          description: Matching Stripe Checkout session ID that can be exchanged for a premium access token.
        sessionId:
          type: string
          description: Alias for checkoutSessionId.
        network:
          type: string
          enum: [visa, mastercard, amex, discover]
        issuer:
          type: string
          example: Chase
        cardholderName:
          type: string
          example: Jane Doe
        addressLine1:
          type: string
          example: 123 Main St
        addressLine2:
          type: string
          example: Miami, FL 33101
        phone:
          type: string
          example: "3055551212"
        merchantName:
          type: string
          example: Example Merchant
        merchantVertical:
          type: string
          example: subscription
        transactionAmount:
          oneOf:
            - type: number
            - type: string
          example: $1,234.56
          description: Flexible amount input. The API normalizes common U.S. and international display formats.
        transactionDate:
          type: string
          example: 03/01/2026
          description: Flexible date input. The API normalizes common slash, dash, dot, and month-name formats.
        description:
          type: string
          description: Primary dispute summary used for reason-code matching.
          example: I cancelled before renewal but the merchant chargd me anyway.
        evidenceSummary:
          type: string
          example: I have cancellation emails and screenshots of the account page.
        evidenceItems:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
        timelineItems:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
        merchantRebuttalConcerns:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
        desiredOutcome:
          type: string
          example: Reverse the charge and confirm the billing correction in writing.
        tone:
          type: string
          enum: [formal, assertive, concise, empathetic]
          example: formal
        lengthPreference:
          type: string
          enum: [short, standard, detailed]
          example: detailed
        outputFormat:
          type: string
          enum: [pdf, docx, word-compatible-rtf, text]
          example: pdf
        async:
          type: boolean
          description: When true on supported premium endpoints, the server returns a queued job instead of waiting for completion.
          example: true
        redactionMode:
          type: string
          enum: [none, standard, strict]
          description: Optional privacy mode for generated report artifacts.
          example: standard
        includeRedactedVersion:
          type: boolean
          description: When true, the report-document endpoint returns both a full copy and a redacted copy.
          example: true
      required: [description]
...
    Exhibit:
      type: object
      properties:
        label:
          type: string
          example: Exhibit A
        title:
          type: string
        description:
          type: string
        filename:
          type: string
          nullable: true
        status:
          type: string
          enum: [provided, suggested]
      required: [label, title, description, status]
    ExhibitPacket:
      type: object
      properties:
        exhibits:
          type: array
          items:
            $ref: "#/components/schemas/Exhibit"
        exhibitIndex:
          type: array
          items:
            type: string
        providedCount:
          type: integer
        suggestedCount:
          type: integer
        recommendedPackageOrder:
          type: array
          items:
            type: string
      required: [exhibits, exhibitIndex, providedCount, suggestedCount, recommendedPackageOrder]
    ExtractedEvidenceFile:
      type: object
      properties:
        filename:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        sourceType:
          type: string
          enum: [text, pdf, image, binary]
        extractionMode:
          type: string
          example: openai-ocr
        ocrApplied:
          type: boolean
        documentType:
          type: string
        summary:
          type: string
        extractedText:
          type: string
        evidenceItems:
          type: array
          items:
            type: string
        timelineItems:
          type: array
          items:
            type: string
        rebuttalConcerns:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
      required:
        - filename
        - mimeType
        - size
        - sourceType
        - extractionMode
        - ocrApplied
        - documentType
        - summary
        - extractedText
        - evidenceItems
        - timelineItems
        - rebuttalConcerns
        - warnings
    EvidenceExtractionCombined:
      type: object
      properties:
        summary:
          type: string
        evidenceItems:
          type: array
          items:
            type: string
        timelineItems:
          type: array
          items:
            type: string
        rebuttalConcerns:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        recommendedNextSteps:
          type: array
          items:
            type: string
      required:
        - summary
        - evidenceItems
        - timelineItems
        - rebuttalConcerns
        - warnings
        - recommendedNextSteps
    EvidenceExtractionResponse:
      type: object
      properties:
        email:
          type: string
          format: email
        premiumAccessToken:
          type: string
          nullable: true
        extractedCount:
          type: integer
        files:
          type: array
          items:
            $ref: "#/components/schemas/ExtractedEvidenceFile"
        combined:
          $ref: "#/components/schemas/EvidenceExtractionCombined"
        exhibitPacket:
          $ref: "#/components/schemas/ExhibitPacket"
        storedFiles:
          type: array
          items:
            $ref: "#/components/schemas/StoredFileSummary"
        caseFile:
          $ref: "#/components/schemas/CaseFileSummary"
        caseVersion:
          $ref: "#/components/schemas/CaseVersion"
      required: [email, extractedCount, files, combined, exhibitPacket, storedFiles, caseFile, caseVersion]
...
    StoredFileSummary:
      type: object
      properties:
        fileId:
          type: string
        originalFilename:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
        sha256:
          type: string
      required: [fileId, originalFilename, contentType, sizeBytes, sha256]
...
    SubmissionPlanStep:
      type: object
      properties:
        order:
          type: integer
        title:
          type: string
        description:
          type: string
      required: [order, title, description]
    SubmissionPlan:
      type: object
      properties:
        recommendedPackageOrder:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/SubmissionPlanStep"
        qualityChecks:
          type: array
          items:
            type: string
      required: [recommendedPackageOrder, steps, qualityChecks]
    CaseFileSummary:
      type: object
      properties:
        caseId:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        latestStage:
          type: string
        versionCount:
          type: integer
        merchantName:
          type: string
        issuer:
          type: string
        network:
          type: string
          nullable: true
        reasonCode:
          type: string
          nullable: true
        confidence:
          type: integer
          nullable: true
        merchantVertical:
          type: string
          nullable: true
        latestArtifactFileIds:
          type: array
          items:
            type: string
        latestArtifactFormats:
          type: array
          items:
            type: string
        latestArtifactUrls:
          type: array
          items:
            type: string
      required:
        - caseId
        - email
        - createdAt
        - updatedAt
        - latestStage
        - versionCount
        - latestArtifactFileIds
        - latestArtifactFormats
        - latestArtifactUrls
    CaseVersion:
      type: object
      properties:
        versionId:
          type: string
        createdAt:
          type: string
          format: date-time
        stage:
          type: string
      required: [versionId, createdAt, stage]
...
    JobAcceptedResponse:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [pending, running]
        email:
          type: string
          format: email
        premiumAccessToken:
          type: string
          nullable: true
        requestId:
          type: string
          nullable: true
        storedFiles:
          type: array
          items:
            $ref: "#/components/schemas/StoredFileSummary"
      required: [jobId, status, email]
...
    DenialResponse:
      type: object
      properties:
        email:
          type: string
          format: email
        premiumAccessToken:
          type: string
          nullable: true
        caseId:
          oneOf:
            - type: string
            - type: "null"
        reasonCode:
          type: string
          nullable: true
        network:
          type: string
        denialSummary:
          oneOf:
            - type: string
            - type: "null"
        counterLetter:
          type: string
        rebuttalTargets:
          type: array
          items:
            type: string
        counterStrategy:
          type: array
          items:
            type: string
        additionalEvidenceRequests:
          type: array
          items:
            type: string
        submissionPlan:
          $ref: "#/components/schemas/SubmissionPlan"
        reviewFlags:
          type: array
          items:
            type: string
        caseFile:
          $ref: "#/components/schemas/CaseFileSummary"
        caseVersion:
          $ref: "#/components/schemas/CaseVersion"
      required: [email, network, counterLetter, rebuttalTargets, counterStrategy, additionalEvidenceRequests, submissionPlan, reviewFlags, caseFile, caseVersion]
...
    UpgradeRequiredResponse:
      type: object
      properties:
        upgradeRequired:
          type: boolean
          example: true
        checkoutUrl:
          type: string
          format: uri
          example: https://checkout.stripe.com/pay/cs_test_123
        checkoutSessionId:
          type: string
          example: cs_test_123
        message:
          type: string
          example: Premium access requires a one-time $6.99 payment.
      required: [upgradeRequired, checkoutUrl, checkoutSessionId, message]